requirements.txt
Adafruit-CharLCD==1.1.1
Adafruit-GPIO==1.0.4
Adafruit-PureIO==0.2.3
aniso8601==8.0.0
Click==7.0
enum-compat==0.0.3
Flask==1.1.1
Flask-RESTful==0.3.7
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pytz==2019.3
RPi.GPIO==0.7.0
schedule==0.6.0
six==1.13.0
spidev==3.4
Werkzeug==0.16.0
yeelight==0.5.0
